LayoutTests/fast/viewport/ios/shrink-to-fit-content-before-document-load.html

<!DOCTYPE html> <!-- webkit-test-runner [ shouldIgnoreMetaViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<style>
body, html {
    margin: 0;
    width: 100%;
    height: 100%;
}

#bar {
    width: 1280px;
    height: 100px;
    background: linear-gradient(to right, red 0%, green 50%, blue 100%);
}
</style>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<script>
description("Verifies that the shrink-to-fit heuristic is given a chance to run before document load finishes. This test requires WebKitTestRunner.");

addEventListener("DOMContentLoaded", () => {
    shouldBe("innerWidth", "1280");
    shouldBe("visualViewport.scale.toFixed(2)", "(screen.width / 1280).toFixed(2)");
});
</script>
</head>
<body>
<div id="bar"></div>
<div id="description"></div>
<div class="square"></div>
<script>
document.scrollingElement.scrollTo(0, 1);
document.scrollingElement.scrollTop;
</script>
</body>
</html>
