LayoutTests/fast/viewport/ios/shrink-to-fit-content-before-document-load-expected.txt

Verifies that the shrink-to-fit heuristic is given a chance to run before document load finishes. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS innerWidth is 1280
PASS visualViewport.scale.toFixed(2) is (screen.width / 1280).toFixed(2)
PASS successfullyParsed is true

TEST COMPLETE
